Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
-
-int main(){
+float main(){
 float a;
 scanf("%f",&a);
 printf("You entered: %f",a);
